implemented REST API, adjusted frontend part with it

+++ src/main/java/com/levi9/domain/Verification.java
@@ -7,7 +7,6 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.OneToOne;
@@ -22,7 +21,7 @@ public class Verification {
 
     @Id
     @GeneratedValue
-    private Integer id;
+    private Long id;
 
     private String token;
 
@@ -74,11 +73,11 @@ public class Verification {
         this.expirationDate = expirationDate;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(final Integer id) {
+    public void setId(final Long id) {
         this.id = id;
     }
 }
