+++ src/main/java/com/levi9/repositories/VerificationRepository.java
@@ -5,7 +5,7 @@ import com.levi9.domain.Verification;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VerificationRepository extends JpaRepository<Verification, Integer> {
+public interface VerificationRepository extends JpaRepository<Verification, Long> {
 
     Verification findByToken(String token);
 }
